added jwt on growing lists

+++ src/api/GrowingLists/deleteGrowingLists.js
@@ -2,7 +2,9 @@ const growingLists = require("../../models/GrowingList");
 
 const deleteGrowingList = async (req, res, next) => {
   try {
-
+    if (req.user.email !== req.query.email) {
+      return res.status(403).send({ message: "Forbidden Access" });
+    }
     const { id } = req.params;
     const camp = await growingLists.findByIdAndDelete(id);
     if (!camp) {
